refactor: code moving
