+++ node/src/node/slave/state/spawn.cpp
@@ -21,7 +21,6 @@
 #include "cocaine/detail/service/node/slave/control.hpp"
 #include "cocaine/detail/service/node/slave/spawn_handle.hpp"
 #include "cocaine/detail/service/node/slave/state/handshaking.hpp"
-#include "cocaine/detail/service/node/util.hpp"
 
 namespace cocaine {
 namespace detail {
