+++ node/src/util/bound.hpp
@@ -1,12 +1,14 @@
 #pragma once
 
 namespace cocaine {
-namespace detail {
+namespace util {
 
 template<typename T>
-constexpr const T& bound(const T& min, const T& value, const T& max) {
+constexpr
+const T&
+bound(const T& min, const T& value, const T& max) {
     return std::max(min, std::min(value, max));
 }
 
-}  // namespace detail
+}  // namespace util
 }  // namespace cocaine
